+++ server/tsuserver.py
@@ -45,8 +45,8 @@ class TsuserverDR:
         self.release = 4
         self.major_version = 2
         self.minor_version = 4
-        self.segment_version = 'a1'
-        self.internal_version = 'm200410a'
+        self.segment_version = 'a2'
+        self.internal_version = 'm200411a'
         version_string = self.get_version_string()
         self.software = 'TsuserverDR {}'.format(version_string)
         self.version = 'TsuserverDR {} ({})'.format(version_string, self.internal_version)
@@ -245,6 +245,12 @@ class TsuserverDR:
         with Constants.fopen('config/config.yaml', 'r', encoding='utf-8') as cfg:
             self.config = Constants.yaml_load(cfg)
             self.config['motd'] = self.config['motd'].replace('\\n', ' \n')
+            self.config['passwords'] = []
+            passwords = ['modpass', 'cmpass', 'gmpass']
+            for i in range(1, 8):
+                passwords.append('gmpass{}'.format(i))
+            for password in passwords:
+                self.config['passwords'].append(self.config[password])
 
         for i in range(1, 8):
             daily_gmpass = 'gmpass{}'.format(i)
@@ -263,6 +269,7 @@ class TsuserverDR:
                              'blackout_background': 'Blackout_HD',
                              'default_area_description': 'No description.',
                              'party_lights_timeout': 10,
+                             'show_ms2-prober': True,
                              'showname_max_length': 30,
                              'sneak_handicap': 5,
                              'spectator_name': 'SPECTATOR',